project/src/requirements.txt
category==0.9.10
click==6.7
django-debug-toolbar==1.9.1
ipython==5.5.0
Django==1.11.11
mysqlclient==1.3.12
